src/main/java/com/fraud/system/services/AuthorizationService.java
package com.fraud.system.services;

import com.fraud.system.domain.user.User;
import com.fraud.system.exceptions.CustomException;
import com.fraud.system.repositories.UserRepository;
import lombok.AllArgsConstructor;
import lombok.NoArgsConstructor;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.stereotype.Service;

@Service
@AllArgsConstructor
public class AuthorizationService implements UserDetailsService {

    @Autowired
    UserRepository repository;

    @Override
    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
        User user = repository.findByLogin(username);

        if (user == null) {
            throw new CustomException(HttpStatus.BAD_REQUEST.value(), "User not found with username: " + username);
        }

        return user;
    }
}
